[M] Rename RandomQuote to RandomSpan
